// BookAPI/ClientApp/src/app/models/FrmLote.ts
export interface FrmLote {
    lote: string;
    item: string;
    tipoitem: string;
    tipolote: number;
    loteorigem: string;
    tipoloteorigem: number;
    temlotefilhoorigem: number;
    datarecebimento: string;
    dataatualizacao: string;
    datafimlote: string;
    datafimreceb: string;
    etiquetaimpressa: number;
    num_aviso_rec: number;
    notafiscal: number;
    codcliente: string;
    nomecliente: string;
    codfornec: string;
    fornecedor: string;
    pesonf: number;
    pesobruto: number;
    pesotara: number;
    pesoliquido: number;
    pesodescontado: number;
    saldo: number;
    envioajustelogix: string;
    flagliberacao: number;
    responsavelbal: string;
    RespReceb: string;
    observacao: string;
    statusreceb: number;
    statuslote: number;
    flagfifoliberado: number;
    SaldoSilo: number;
    Saldo_Reservado: number;
    DescTipoLote: string;
    DescItem: string;
    StLote: string;
}